+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [version 1.7.3] 2024-03-04
 
+- Upgrade packages: [file_picker: 8.0.0+1], [sqlite3: 2.4.2], [dart_pdf_reader: ^1.0.0]
 - Better Sorting fallback
 - Set API Key directly from the Stock Chart